Build the supra-transition matrix for a multilayer network

Given a supra-adjacency matrix, build supra-transition the matrix corresponding to different type of random walks (discrete time Markov chians). The types of RW considered here are the same as in Ref.2-Supplementary Material: classical, diffusive, PageRank, physical, physical with relaxation and maximum-entropy.

The maximum entropy random walk (MERW) choses the stochastic matrix which maximises \(H(S)\), so that the walker can explore every walk of the same length with equal probability. Let \(\lambda_N, \phi\) be the leading eigenvalue and corresponding right eigenvector of the adjacency matrix \(A\). Then \(T_{ij} = \frac{A_{ij}}{\lambda_N}\frac{\phi_j}{\phi_i}.\)

BuildSupraTransitionMatrixFromSupraAdjacencyMatrix(
  SupraAdjacencyMatrix,
  Layers,
  Nodes,
  Type = "pagerank"
)

Arguments

SupraAdjacencyMatrix
the supra-adjacency matrix, which contains the information in intra-layer connectivity \(\{W^{\alpha}_{ij}\}\) and on inter-layer coupilng \(D^{\alpha \beta}_{i}\}_i\).

Layers
integer, number of layers

Nodes
integer, number of nodes

Type
the type of the random walk, default "pagerank". Other types: c("classical", "diffusive", "maxent", "physical", "relaxed-physical"). For details on the transition rules see Refs.2-3.

Value

the supra-transition matrix of probabilities for a discrete-time random walk on the given multi-layer network.

References

De Domenico, M., Set al. (2013). Mathematical formulation of multilayer networks. Physical Review X, 3(4), 041022. doi 10.1103/PhysRevX.3.041022

De Domenico, M., et al. (2014). Navigability of interconnected networks under random failures. PNAS, 111(23). doi 10.1073/pnas.1318469111

Bertagnolli, G., & De Domenico, M. (2020). Diffusion Geometry of Multiplex and Interdependent Systems. arxiv preprint arxiv:2006.13032