Random walk transition matrix for an edge-coloured network. Given the node tensor for the edge-coloured multi-layer network the function builds the supra transition matrix for a discrete time RW.

Random walk transition matrix for an edge-coloured network. Given the node tensor for the edge-coloured multi-layer network the function builds the supra transition matrix for a discrete time RW.

BuildTransitionMatrixFromEdgeColoredMatrices(
  NodesTensor,
  Layers,
  Nodes,
  Type = "classical"
)

Arguments

NodesTensor
list of adjacency matrices, expected to be aligned (a node has same index in any layer).

Layers
integer, number of layers

Nodes
integer, number of nodes

Type
the type of the random walk, default "classical". Other types: c("pagerank", "diffusive", "maxent"). For details on the transition rules see Refs.2-3.

Value

a square matrix of dimension Nodes of transition probabilities for a discrete-time random walk on the given edge-coloured network.

References

De Domenico, M., Set al. (2013). Mathematical formulation of multilayer networks. Physical Review X, 3(4), 041022. doi 10.1103/PhysRevX.3.041022

De Domenico, M., et al. (2014). Navigability of interconnected networks under random failures. PNAS, 111(23). doi 10.1073/pnas.1318469111

Bertagnolli, G., & De Domenico, M. (2020). Diffusion Geometry of Multiplex and Interdependent Systems. arxiv preprint arxiv:2006.13032