Plot a multiplex

Plot a multiplex

plot_multiplex(
  g.list,
  layer.colors,
  edge.colors = "auto",
  node.colors = "auto",
  node.size.values = 0.5,
  node.alpha = 1,
  edge.alpha = 1,
  layout = "fr",
  show.legend = TRUE
)

plot.multiplex(g.list, layer.colors, edge.colors = "auto", node.colors = "auto",
   node.size.values = 0.5, node.alpha = 1, edge.alpha = 1, layout = "fr", show.legend = T)

Arguments

g.list
list of networks (representing the layers)

layer.colors
list of colors for the each layer

edge.colors
edge colors. If not provided, use the same color of the layer.

node.colors
node colors. If not provided, use the same color of the layer.

node.size.values
default 0.5,

node.alpha
default 1,

edge.alpha
default 1,

layout
default "fr", see layout_with_fr for other options.

show.legend
default TRUE

Value

a plot

Methods (by generic)

plot: Old name